Refactor UnusedPrivateMemberSpec
